Fix syntax to avoid rm -rf /

+++ entrypoint-restapi-node-fix.sh
@@ -107,7 +107,7 @@ if [ -e "$DB_DIR/flush_lock" ] ; then
   log "Flush_lock was found!, trying restoring DB from snapshot"
   if [ -d "$DB_DIR/../database-snapshot" ] ; then
     log "Cleaning up database directory $DB_DIR"
-    rm -rf $DB_DIR/*
+    rm -rf ${DB_DIR:?}/*
     log "Restoring database from database snaphost"
     cp -R --reflink $DB_DIR/../database-snapshot/* $DB_DIR
   else
